launcher refactoring - ArchiveHandler
- first working version of ArchiveHandler using Kodis builtin archive support
